+++ apps/tx-builder/src/hardhat/deploy/deploy_matrix_test_contract.js
@@ -1,0 +1,13 @@
+const deploy = async function (hre) {
+  const { deployments, getNamedAccounts } = hre
+  const { deployer } = await getNamedAccounts()
+  const { deploy } = deployments
+
+  await deploy('MatrixTypesTestContract', {
+    from: deployer,
+    args: [],
+    log: true,
+  })
+}
+
+module.exports = deploy