+++ apps/tx-builder/src/lib/storage.ts
@@ -82,7 +82,7 @@ const importBatch = async (file: File): Promise<BatchFile> => {
   return new Promise(resolve => {
     const reader = new FileReader()
     reader.readAsText(file)
-    reader.onload = async () => {
+    reader.onload = () => {
       const batchFile: BatchFile = JSON.parse(reader.result as string)
       resolve(batchFile)
     }
@@ -91,7 +91,7 @@ const importBatch = async (file: File): Promise<BatchFile> => {
 
 const uuidv4 = () => {
   return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function (c) {
-    var r = (Math.random() * 16) | 0,
+    const r = (Math.random() * 16) | 0,
       v = c === 'x' ? r : (r & 0x3) | 0x8
     return v.toString(16)
   })