+++ apps/tx-builder/src/pages/ReviewAndConfirm.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useMemo, useState } from 'react'
+import { useEffect, useState } from 'react'
 import {
   Button,
   ButtonLink,
@@ -42,10 +42,14 @@ const ReviewAndConfirm = () => {
     reorderTransactions,
   } = useTransactions()
   const { downloadBatch, saveBatch } = useTransactionLibrary()
-  const rawTransactions = useMemo(() => transactions.map(t => t.raw), [transactions])
   const [showSimulation, setShowSimulation] = useState<boolean>(false)
-  const { simulation, simulateTransaction, simulationRequestStatus, simulationLink } =
-    useSimulation(rawTransactions)
+  const {
+    simulation,
+    simulateTransaction,
+    simulationRequestStatus,
+    simulationLink,
+    simulationSupported,
+  } = useSimulation()
   const navigate = useNavigate()
 
   const clickSimulate = () => {
@@ -116,15 +120,17 @@ const ReviewAndConfirm = () => {
           </Button>
 
           {/* Simulate batch button */}
-          <Button
-            size="md"
-            type="button"
-            variant="contained"
-            color="secondary"
-            onClick={clickSimulate}
-          >
-            Simulate
-          </Button>
+          {simulationSupported && (
+            <Button
+              size="md"
+              type="button"
+              variant="contained"
+              color="secondary"
+              onClick={clickSimulate}
+            >
+              Simulate
+            </Button>
+          )}
         </ButtonsWrapper>
 
         {/* Simulation statuses */}
@@ -136,6 +142,12 @@ const ReviewAndConfirm = () => {
               color="inputFilled"
               onClick={closeSimulation}
             ></StyledButton>
+            {simulationRequestStatus === FETCH_STATUS.ERROR && (
+              <Text color="error" size="lg">
+                An unexpected error occurred during simulation.
+              </Text>
+            )}
+
             {simulationRequestStatus === FETCH_STATUS.LOADING && (
               <>
                 <Loader size="xs" />
@@ -248,8 +260,7 @@ const SimulationContainer = styled(Card)`
 
 const Wrapper = styled.main`
   && {
-    padding: 48px;
-    padding-top: 120px;
+    padding: 120px 48px 48px;
     max-width: 650px;
     margin: 0 auto;
   }